--- packages/web/src/components/custom-antd/AutoComplete/AutoComplete.tsx
@@ -1,7 +1,0 @@
-import {AntdCustomStyledAutoComplete, ICustomAutoCompleteProps} from './AutoComplete.styled';
-
-const AutoComplete: React.FC<ICustomAutoCompleteProps> = props => {
-  return <AntdCustomStyledAutoComplete {...props} />;
-};
-
-export default AutoComplete;
