--- packages/web/src/components/custom-antd/AutoComplete/index.ts
@@ -1,1 +1,0 @@
-export {default} from './AutoComplete';
