+++ packages/web/src/components/custom-antd/index.ts
@@ -11,4 +11,3 @@ export {default as FullWidthSpace} from './FullWidthSpace';
 export {default as FormRow} from './Form/FormRow';
 export {default as FormItemLabel} from './Form/FormItem/FormItemLabel';
 export {FormIconWrapper} from './Form/Form.styled';
-export {default as AutoComplete} from './AutoComplete';
